+++ components/TransactionTable.tsx
@@ -1,14 +1,11 @@
 import React, { useEffect, useMemo } from "react";
 import { useState, useRef } from "react";
 import {
-  ArrowDown,
-  ArrowUp,
   Check,
   ChevronDown,
   ChevronRight,
   FileText,
   Layers,
-  ListFilter,
   Paperclip,
   Trash,
   Unlink,
@@ -18,6 +15,7 @@ import { Transaction, SortConfig, Status, STATUSES } from "@/types/transaction";
 import StatusBadge from "./StatusBadge";
 import BulkActions from "./BulkActions";
 import ContextMenu from "./ContextMenu";
+import TransactionTableHeader from "./TransactionTableHeader";
 
 function DriveFileCell({
   fileId,
@@ -357,249 +355,6 @@ const TransactionTable = ({
     return () => document.removeEventListener("mousedown", handler);
   }, [openFilterCol]);
 
-  const dropdownBase =
-    "absolute top-full left-0 mt-0.5 bg-white dark:bg-[#1b1b1b] border border-gray-200 dark:border-gray-700 rounded shadow-lg z-50 py-1 font-normal normal-case tracking-normal text-left select-auto";
-
-  const renderFilterDropdown = (
-    col: "category" | "status" | "source",
-    options: string[],
-    showNone = false,
-  ) => {
-    const active = columnFilters[col];
-    return (
-      <div
-        ref={filterRef}
-        className={`${dropdownBase} min-w-44`}
-        onMouseDown={(e) => e.stopPropagation()}
-      >
-        <div className="px-3 py-1.5 flex items-center justify-between border-b border-gray-100 dark:border-gray-800">
-          <span className="text-[11px] text-gray-400 dark:text-gray-500">
-            Filter
-          </span>
-          {active.length > 0 && (
-            <button
-              onClick={() => onFilterChange(col, [])}
-              className="text-[11px] text-gray-600 hover:text-gray-900 dark:text-gray-400 dark:hover:text-foreground"
-            >
-              Clear
-            </button>
-          )}
-        </div>
-        <div className="max-h-52 overflow-y-auto">
-          {options.map((opt) => (
-            <label
-              key={opt}
-              className="flex items-center gap-2 px-3 py-1.5 text-[12px] cursor-pointer hover:bg-gray-50 dark:hover:bg-[#282828] text-gray-900 dark:text-foreground"
-            >
-              <input
-                type="checkbox"
-                checked={active.includes(opt)}
-                onChange={() => {
-                  const next = active.includes(opt)
-                    ? active.filter((v) => v !== opt)
-                    : [...active, opt];
-                  onFilterChange(col, next);
-                }}
-                className="w-3 h-3 accent-gray-700 dark:accent-gray-300 cursor-pointer"
-              />
-              {opt}
-            </label>
-          ))}
-          {showNone && (
-            <label className="flex items-center gap-2 px-3 py-1.5 text-[12px] cursor-pointer hover:bg-gray-50 dark:hover:bg-[#282828] text-gray-400 dark:text-gray-500">
-              <input
-                type="checkbox"
-                checked={active.includes("__none__")}
-                onChange={() => {
-                  const next = active.includes("__none__")
-                    ? active.filter((v) => v !== "__none__")
-                    : [...active, "__none__"];
-                  onFilterChange(col, next);
-                }}
-                className="w-3 h-3 accent-gray-700 dark:accent-gray-300 cursor-pointer"
-              />
-              (None)
-            </label>
-          )}
-        </div>
-      </div>
-    );
-  };
-
-  const dropdownInputClass =
-    "w-full bg-transparent border-b border-gray-200 dark:border-gray-700 focus:border-gray-400 dark:focus:border-gray-500 outline-none text-[12px] text-gray-900 dark:text-foreground py-1 px-0 placeholder-gray-400 dark:placeholder-gray-500 focus:ring-0";
-
-  const renderDescriptionDropdown = () => {
-    const active = textFilters.description !== "";
-    return (
-      <div
-        ref={filterRef}
-        className={`${dropdownBase} w-full`}
-        onMouseDown={(e) => e.stopPropagation()}
-      >
-        <div className="px-3 py-1.5 flex items-center justify-between border-b border-gray-100 dark:border-gray-800">
-          <span className="text-[11px] text-gray-400 dark:text-gray-500">
-            Filter
-          </span>
-          {active && (
-            <button
-              onClick={() => onTextFilterChange("description", "")}
-              className="text-[11px] text-gray-600 hover:text-gray-900 dark:text-gray-400 dark:hover:text-foreground"
-            >
-              Clear
-            </button>
-          )}
-        </div>
-        <div className="px-3 py-2">
-          <input
-            autoFocus
-            type="text"
-            placeholder="Search description..."
-            value={textFilters.description}
-            onChange={(e) => onTextFilterChange("description", e.target.value)}
-            className={dropdownInputClass}
-          />
-        </div>
-      </div>
-    );
-  };
-
-  const renderDateRangeDropdown = () => {
-    const active = textFilters.dateFrom !== "" || textFilters.dateTo !== "";
-    return (
-      <div
-        ref={filterRef}
-        className={`${dropdownBase} w-52`}
-        onMouseDown={(e) => e.stopPropagation()}
-      >
-        <div className="px-3 py-1.5 flex items-center justify-between border-b border-gray-100 dark:border-gray-800">
-          <span className="text-[11px] text-gray-400 dark:text-gray-500">
-            Filter
-          </span>
-          {active && (
-            <button
-              onClick={() => {
-                onTextFilterChange("dateFrom", "");
-                onTextFilterChange("dateTo", "");
-              }}
-              className="text-[11px] text-gray-600 hover:text-gray-900 dark:text-gray-400 dark:hover:text-foreground"
-            >
-              Clear
-            </button>
-          )}
-        </div>
-        <div className="px-3 py-2 flex flex-col gap-2">
-          <div>
-            <p className="text-[10px] text-gray-400 dark:text-gray-500 mb-1 uppercase tracking-wider">
-              From
-            </p>
-            <input
-              type="date"
-              value={textFilters.dateFrom}
-              onChange={(e) => onTextFilterChange("dateFrom", e.target.value)}
-              className={dropdownInputClass}
-            />
-          </div>
-          <div>
-            <p className="text-[10px] text-gray-400 dark:text-gray-500 mb-1 uppercase tracking-wider">
-              To
-            </p>
-            <input
-              type="date"
-              value={textFilters.dateTo}
-              onChange={(e) => onTextFilterChange("dateTo", e.target.value)}
-              className={dropdownInputClass}
-            />
-          </div>
-        </div>
-      </div>
-    );
-  };
-
-  const renderAmountRangeDropdown = () => {
-    const active = textFilters.amountMin !== "" || textFilters.amountMax !== "";
-    return (
-      <div
-        ref={filterRef}
-        className={`${dropdownBase} w-44`}
-        onMouseDown={(e) => e.stopPropagation()}
-      >
-        <div className="px-3 py-1.5 flex items-center justify-between border-b border-gray-100 dark:border-gray-800">
-          <span className="text-[11px] text-gray-400 dark:text-gray-500">
-            Filter
-          </span>
-          {active && (
-            <button
-              onClick={() => {
-                onTextFilterChange("amountMin", "");
-                onTextFilterChange("amountMax", "");
-              }}
-              className="text-[11px] text-gray-600 hover:text-gray-900 dark:text-gray-400 dark:hover:text-foreground"
-            >
-              Clear
-            </button>
-          )}
-        </div>
-        <div className="px-3 py-2 flex flex-col gap-2">
-          <div>
-            <p className="text-[10px] text-gray-400 dark:text-gray-500 mb-1 uppercase tracking-wider">
-              Min
-            </p>
-            <input
-              type="number"
-              step="0.01"
-              placeholder="e.g. -100"
-              value={textFilters.amountMin}
-              onChange={(e) => onTextFilterChange("amountMin", e.target.value)}
-              className={dropdownInputClass}
-            />
-          </div>
-          <div>
-            <p className="text-[10px] text-gray-400 dark:text-gray-500 mb-1 uppercase tracking-wider">
-              Max
-            </p>
-            <input
-              type="number"
-              step="0.01"
-              placeholder="e.g. 500"
-              value={textFilters.amountMax}
-              onChange={(e) => onTextFilterChange("amountMax", e.target.value)}
-              className={dropdownInputClass}
-            />
-          </div>
-        </div>
-        <div className="px-3 py-2 border-t border-gray-100 dark:border-gray-800 flex items-center justify-between">
-          <span className="text-[11px] text-gray-500 dark:text-gray-400">
-            Sort by abs value
-          </span>
-          <button
-            onMouseDown={(e) => e.preventDefault()}
-            onClick={() => onToggleAbsSort()}
-            className={`relative w-7 h-4 rounded-full transition-colors ${
-              absValue ? "bg-blue-500" : "bg-gray-200 dark:bg-gray-700"
-            }`}
-          >
-            <span
-              className={`absolute top-0.5 left-0 w-3 h-3 rounded-full bg-white transition-transform ${
-                absValue ? "translate-x-3.5" : "translate-x-0.5"
-              }`}
-            />
-          </button>
-        </div>
-      </div>
-    );
-  };
-
-  const selectableIds = useMemo(
-    () => transactions.map((tx) => tx.id),
-    [transactions],
-  );
-
-  const allSelected =
-    selectableIds.length > 0 &&
-    selectableIds.every((id) => selectedIds.has(id));
-  const someSelected = selectableIds.some((id) => selectedIds.has(id));
-
   const canGroup = useMemo(() => {
     const selected = [...selectedIds.values()];
     if (selected.length < 1) return false;
@@ -743,19 +498,9 @@ const TransactionTable = ({
     );
   };
 
-  const renderSortIcon = (key: keyof Transaction) => {
-    if (!sortConfig || sortConfig.key !== key) return null;
-    return sortConfig.direction === "asc" ? (
-      <ArrowUp className="w-3 h-3 ml-1 inline-block text-gray-400 dark:text-gray-500" />
-    ) : (
-      <ArrowDown className="w-3 h-3 ml-1 inline-block text-gray-400 dark:text-gray-500" />
-    );
-  };
-
   const isEditing = (id: string, field: EditableFields) =>
     editingCell?.id === id && editingCell?.field === field;
 
-  const thClass = `h-9 px-4 font-normal text-[11px] uppercase tracking-wider text-gray-600 dark:text-gray-400 border-b border-r border-gray-200 dark:border-gray-700 select-none`;
   const tdClass = `h-9 px-4 text-[13px] border-b border-r border-gray-100 dark:border-gray-800 text-gray-600 dark:text-gray-400 whitespace-nowrap`;
   const addInputClass = `w-full bg-transparent border-0 border-b border-transparent hover:border-gray-200 dark:hover:border-gray-700 focus:ring-0 p-1 text-[13px] text-gray-900 dark:text-foreground placeholder-gray-400 dark:placeholder-gray-500 transition-colors outline-none`;
   const editInputClass = `w-full bg-transparent border-0 outline-none text-[13px] text-gray-900 dark:text-foreground p-0 m-0 focus:ring-0 caret-gray-400 dark:caret-gray-500`;
@@ -1118,170 +863,22 @@ const TransactionTable = ({
             <col className="w-14" />
           </colgroup>
           {/* Header */}
-          <thead className="sticky top-0 bg-white dark:bg-[#131314] z-2">
-            <tr>
-              <th className={`${thClass} w-8`}>
-                {selectableIds.length > 0 && (
-                  <label className="flex items-center justify-center cursor-pointer">
-                    <input
-                      type="checkbox"
-                      checked={allSelected}
-                      ref={(el) => {
-                        if (el) el.indeterminate = someSelected && !allSelected;
-                      }}
-                      onChange={() => {
-                        if (allSelected) {
-                          for (const tx of transactions) {
-                            onToggleSelect(tx);
-                          }
-                        } else {
-                          onSelectAll(transactions);
-                        }
-                      }}
-                      className="w-3.5 h-3.5 accent-gray-700 dark:accent-gray-300 cursor-pointer"
-                    />
-                  </label>
-                )}
-              </th>
-              <th className={`${thClass} relative w-34`}>
-                <div className="flex items-center justify-between gap-1">
-                  <span
-                    className="flex items-center cursor-pointer hover:text-gray-900 dark:hover:text-foreground transition-colors"
-                    onClick={() => onSort("date")}
-                  >
-                    Date {renderSortIcon("date")}
-                  </span>
-                  <button
-                    onClick={(e) => {
-                      e.stopPropagation();
-                      setOpenFilterCol(
-                        openFilterCol === "date" ? null : "date",
-                      );
-                    }}
-                    className={`p-0.5 rounded transition-colors ${textFilters.dateFrom || textFilters.dateTo ? "text-blue-500 dark:text-blue-400" : "text-gray-300 hover:text-gray-500 dark:text-gray-600 dark:hover:text-gray-400"}`}
-                  >
-                    <ListFilter className="w-3 h-3" />
-                  </button>
-                </div>
-                {openFilterCol === "date" && renderDateRangeDropdown()}
-              </th>
-              <th className={`${thClass} relative min-w-64`}>
-                <div className="flex items-center justify-between gap-1">
-                  <span
-                    className="flex items-center cursor-pointer hover:text-gray-900 dark:hover:text-foreground transition-colors"
-                    onClick={() => onSort("description")}
-                  >
-                    Description {renderSortIcon("description")}
-                  </span>
-                  <button
-                    onClick={(e) => {
-                      e.stopPropagation();
-                      setOpenFilterCol(
-                        openFilterCol === "description" ? null : "description",
-                      );
-                    }}
-                    className={`p-0.5 rounded transition-colors ${textFilters.description ? "text-blue-500 dark:text-blue-400" : "text-gray-300 hover:text-gray-500 dark:text-gray-600 dark:hover:text-gray-400"}`}
-                  >
-                    <ListFilter className="w-3 h-3" />
-                  </button>
-                </div>
-                {openFilterCol === "description" && renderDescriptionDropdown()}
-              </th>
-              <th className={`${thClass} relative w-38`}>
-                <div className="flex items-center justify-between gap-1">
-                  <span
-                    className="flex items-center cursor-pointer hover:text-gray-900 dark:hover:text-foreground transition-colors"
-                    onClick={() => onSort("category")}
-                  >
-                    Category {renderSortIcon("category")}
-                  </span>
-                  <button
-                    onClick={(e) => {
-                      e.stopPropagation();
-                      setOpenFilterCol(
-                        openFilterCol === "category" ? null : "category",
-                      );
-                    }}
-                    className={`p-0.5 rounded transition-colors ${columnFilters.category.length > 0 ? "text-blue-500 dark:text-blue-400" : "text-gray-300 hover:text-gray-500 dark:text-gray-600 dark:hover:text-gray-400"}`}
-                  >
-                    <ListFilter className="w-3 h-3" />
-                  </button>
-                </div>
-                {openFilterCol === "category" &&
-                  renderFilterDropdown("category", allCategories, true)}
-              </th>
-              <th className={`${thClass} relative w-32`}>
-                <div className="flex items-center justify-between gap-1">
-                  <span
-                    className="flex items-center cursor-pointer hover:text-gray-900 dark:hover:text-foreground transition-colors"
-                    onClick={() => onSort("amount")}
-                  >
-                    Amount {renderSortIcon("amount")}
-                  </span>
-                  <button
-                    onClick={(e) => {
-                      e.stopPropagation();
-                      setOpenFilterCol(
-                        openFilterCol === "amount" ? null : "amount",
-                      );
-                    }}
-                    className={`p-0.5 rounded transition-colors ${textFilters.amountMin || textFilters.amountMax || absValue ? "text-blue-500 dark:text-blue-400" : "text-gray-300 hover:text-gray-500 dark:text-gray-600 dark:hover:text-gray-400"}`}
-                  >
-                    <ListFilter className="w-3 h-3" />
-                  </button>
-                </div>
-                {openFilterCol === "amount" && renderAmountRangeDropdown()}
-              </th>
-              <th className={`${thClass} relative w-32`}>
-                <div className="flex items-center justify-between gap-1">
-                  <span
-                    className="flex items-center cursor-pointer hover:text-gray-900 dark:hover:text-foreground transition-colors"
-                    onClick={() => onSort("status")}
-                  >
-                    Status {renderSortIcon("status")}
-                  </span>
-                  <button
-                    onClick={(e) => {
-                      e.stopPropagation();
-                      setOpenFilterCol(
-                        openFilterCol === "status" ? null : "status",
-                      );
-                    }}
-                    className={`p-0.5 rounded transition-colors ${columnFilters.status.length > 0 ? "text-blue-500 dark:text-blue-400" : "text-gray-300 hover:text-gray-500 dark:text-gray-600 dark:hover:text-gray-400"}`}
-                  >
-                    <ListFilter className="w-3 h-3" />
-                  </button>
-                </div>
-                {openFilterCol === "status" &&
-                  renderFilterDropdown("status", STATUSES)}
-              </th>
-              <th className={`${thClass} relative w-38`}>
-                <div className="flex items-center justify-between gap-1">
-                  <span
-                    className="flex items-center cursor-pointer hover:text-gray-900 dark:hover:text-foreground transition-colors"
-                    onClick={() => onSort("source")}
-                  >
-                    Source {renderSortIcon("source")}
-                  </span>
-                  <button
-                    onClick={(e) => {
-                      e.stopPropagation();
-                      setOpenFilterCol(
-                        openFilterCol === "source" ? null : "source",
-                      );
-                    }}
-                    className={`p-0.5 rounded transition-colors ${columnFilters.source.length > 0 ? "text-blue-500 dark:text-blue-400" : "text-gray-300 hover:text-gray-500 dark:text-gray-600 dark:hover:text-gray-400"}`}
-                  >
-                    <ListFilter className="w-3 h-3" />
-                  </button>
-                </div>
-                {openFilterCol === "source" &&
-                  renderFilterDropdown("source", allSources, true)}
-              </th>
-              <th className={`${thClass} w-14`}>File</th>
-              <th className={`${thClass} w-14`} />
-            </tr>
-          </thead>
+          <TransactionTableHeader
+            transactions={transactions}
+            sortConfig={sortConfig}
+            onSort={onSort}
+            columnFilters={columnFilters}
+            onFilterSelectChange={onFilterChange}
+            textFilters={textFilters}
+            onFilterTextChange={onTextFilterChange}
+            allCategories={allCategories}
+            allSources={allSources}
+            selectedIdsMap={selectedIds}
+            onToggleSelect={onToggleSelect}
+            onSelectAll={onSelectAll}
+            onToggleAbsSort={onToggleAbsSort}
+            absValue={absValue}
+          />
           <tbody>
             {/* Totals Row */}
             {showTotalsRow && (
